+++ include/malloy/charges/charges.hpp
@@ -1,7 +1,10 @@
 #pragma once
 
 // Umbrella header for malloy_charges: charged particles in electric and
-// magnetic fields.
+// magnetic fields, in two and (since M29) three dimensions.
+#include <malloy/charges/charge3d_settings.hpp>
+#include <malloy/charges/charge3d_world.hpp>
 #include <malloy/charges/charge_settings.hpp>
 #include <malloy/charges/charge_world.hpp>
 #include <malloy/charges/charged_particle2d.hpp>
+#include <malloy/charges/charged_particle3d.hpp>